Moves discover into project root
