+++ cosigner/cosignerverifier.go
@@ -7,8 +7,8 @@ import (
 
 	"github.com/lestrrat-go/jwx/v2/jwa"
 	"github.com/lestrrat-go/jwx/v2/jws"
+	"github.com/openpubkey/openpubkey/discover"
 	"github.com/openpubkey/openpubkey/pktoken"
-	"github.com/openpubkey/openpubkey/providers/discover"
 )
 
 type DefaultCosignerVerifier struct {
