+++ providers/github_actions.go
@@ -25,9 +25,9 @@ import (
 	"os"
 
 	"github.com/awnumar/memguard"
+	"github.com/openpubkey/openpubkey/discover"
 	"github.com/openpubkey/openpubkey/pktoken"
 	"github.com/openpubkey/openpubkey/pktoken/clientinstance"
-	"github.com/openpubkey/openpubkey/providers/discover"
 )
 
 const githubIssuer = "https://token.actions.githubusercontent.com"
